requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
sqlalchemy
psycopg2-binary
python-dotenv
openai
